+++ src/main/java/com/ringoid/api/lmm/Matches.java
@@ -155,7 +155,7 @@ public class Matches {
         parameters.put("sourceUserId", request.getUserId());
         parameters.put("hiddenUserStatus", UserStatus.HIDDEN.getValue());
 
-        int lastActionTime = Utils.lastActionTime(parameters, driver);
+        long lastActionTime = Utils.lastActionTime(parameters, driver);
         LMMResponse response = new LMMResponse();
         response.setLastActionTime(lastActionTime);
