+++ src/main/java/com/ringoid/api/lmm/Messages.java
@@ -73,7 +73,7 @@ public class Messages {
         parameters.put("sourceUserId", request.getUserId());
         parameters.put("hiddenUserStatus", UserStatus.HIDDEN.getValue());
 
-        int lastActionTime = Utils.lastActionTime(parameters, driver);
+        long lastActionTime = Utils.lastActionTime(parameters, driver);
         LMMResponse response = new LMMResponse();
         response.setLastActionTime(lastActionTime);
